requirements.txt
Django==5.2.9
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.9.0
django-filter==25.2
Pillow==12.1.1
psycopg2-binary==2.9.11
